File structure fix & better ux experience
