--- apps/web/src/components/shadcn/nav-secondary.tsx
@@ -1,44 +1,0 @@
-import * as React from "react";
-import { type LucideIcon } from "lucide-react";
-
-import {
-	SidebarGroup,
-	SidebarGroupContent,
-	SidebarMenu,
-	SidebarMenuButton,
-	SidebarMenuItem,
-} from "@/components/shadcn/ui/sidebar";
-
-export function NavSecondary({
-	items,
-	...props
-}: {
-	items: {
-		title: string;
-		url: string;
-		icon: LucideIcon;
-	}[];
-} & React.ComponentPropsWithoutRef<typeof SidebarGroup>) {
-	return (
-		<SidebarGroup {...props}>
-			<SidebarGroupContent>
-				<SidebarMenu>
-					{items.map((item) => (
-						<SidebarMenuItem key={item.title}>
-							<SidebarMenuButton
-								asChild
-								size="lg"
-								className="text-base [&_svg]:h-5 [&_svg]:w-5"
-							>
-								<a href={item.url}>
-									<item.icon />
-									<span>{item.title}</span>
-								</a>
-							</SidebarMenuButton>
-						</SidebarMenuItem>
-					))}
-				</SidebarMenu>
-			</SidebarGroupContent>
-		</SidebarGroup>
-	);
-}
